requirements.txt
azure-core==1.19.0
azure-storage-blob==12.9.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cryptography==35.0.0
dnspython==1.16.0
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msrest==0.6.21
oauthlib==3.1.1
pycparser==2.20
pymongo==3.12.0
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.1
